% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_95.R
\name{CD_95}
\alias{CD_95}
\title{Signed and Signed Rank Tests of \insertCite{dufour1995exact;textual}{Rdpack}}
\usage{
CD_95(y, x, level = 0.05, p = 0.5, W = FALSE, SR = FALSE)
}
\arguments{
\item{y}{the vector of dependent variables}

\item{x}{the \eqn{n\times 1} vector or \eqn{n\times k} matrix of regressors.}

\item{level}{the level of the test -i.e. \eqn{\alpha}. Default value is 0.05.}

\item{p}{is the success probability of the binomial distribution for each trial. Default value is 0.5.}

\item{W}{includes the Wilcoxon signed rank test variate when set to TRUE.}

\item{SR}{includes signed rank test variate proposed by \insertCite{dufour1995exact;textual}{Rdpack} when set to TRUE.}
}
\description{
This function provides the test statistic and the critical value for the nonparametric sign-based and signed-rank tests proposed by \insertCite{dufour1995exact;textual}{Rdpack}.
These tests are valid in the presence of a single regressor and no nuisance parameters (e.g. intercept).
}
\examples{
CD_95(y,x)
}
\references{
\insertAllCited{}
}
\keyword{Nonparametric,}
\keyword{exact,}
\keyword{inference,}
\keyword{signs}
