% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_97.R
\name{CD_97}
\alias{CD_97}
\title{Signed and Signed Rank Tests of \insertCite{campbell1997exact;textual}{Rdpack}}
\usage{
CD_97(y, x, level = 0.05, alpha_1 = 0.14 * level, p = 0.5, SRTest = FALSE)
}
\arguments{
\item{y}{the vector of dependent variables}

\item{x}{the \eqn{n\times 1} vector or \eqn{n\times k} matrix of regressors.}

\item{level}{the level of the test -i.e. \eqn{\alpha}. Default value is 0.05.}

\item{alpha_1}{or \eqn{\alpha_1} for empolying the bound-type procedure. Default value is \eqn{0.14\times\alpha}. Note: \eqn{\alpha_1<\alpha}.}

\item{p}{the success probability of the binomial distribution for each trial. Default value is 0.5.}

\item{SR}{includes signed rank test variate proposed by Campbell and Dufour (1995) when set to TRUE.}
}
\description{
This function extends the nonparametric signed and signed rank tests of \insertCite{campbell1997exact;textual}{Rdpack} by allowing the presence of a nuisance parameter.
The nuisance parameter problem is dealt with by employing bound-type procedures. These tests are valid in the presence of a single regressor and a nuisance parameter (e.g. intercept).
}
\examples{
CD_97(y,x)
}
\references{
\insertAllCited{}
}
\keyword{Bonferroni,}
\keyword{Nonparametric,}
\keyword{exact,}
\keyword{inference,}
\keyword{nuisance-parameter}
\keyword{signs,}
