% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POS_Fix.R
\name{POS_Fix}
\alias{POS_Fix}
\title{Point-Optimal Sign-Based Tests of \insertCite{dufour2010exact;textual}{Rdpack}}
\usage{
POS_Fix(y, x, null = c(0, 0), level = 0.05, p = 0.5, B = 10000, ...)
}
\arguments{
\item{y}{the vector of dependent variables}

\item{x}{the vector of regressors}

\item{null}{the null hypothesis}

\item{level}{is the level of the test. Default value is 0.05.}

\item{p}{is the success probability of the binomial distribution for each trial. Default value is 0.5.}

\item{B}{is the number iterations for simulating the}
}
\description{
This function provides the test statistic and the critical values for the nonparametric point-optimal sign-based tests proposed by \insertCite{dufour2010exact;textual}{Rdpack}.
The proposed tests are exact, distribution-free and valid in the presence of nonstandard distributions and heteroskedasticity of unknown form. Moreover, they have the highest power amongst commonly encountered tests that are intended to be
robust against heteroskedasticity.
}
\examples{
POS_Fix(y,x,null=c(0,0))
}
\references{
\insertAllCited{}
}
\keyword{(2010)}
\keyword{Dufour}
\keyword{Nonparametric}
\keyword{Sign-tests}
\keyword{Taamouti}
\keyword{and}
\keyword{point-optimal}
