% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SB_Dep.R
\name{SB_Dep}
\alias{SB_Dep}
\title{Signed and HAC-Corrected Signed Tests of \insertCite{coudin2009finite;textual}{Rdpack}}
\usage{
SB_Dep(
  y,
  x,
  null = c(0, 0),
  level = 0.05,
  p = 0.5,
  Dpndnt = FALSE,
  B = 10000,
  bwidth
)
}
\arguments{
\item{y}{the vector of dependent variables.}

\item{x}{the \eqn{n\times 1} vector or \eqn{n\times k} matrix of regressors.}

\item{null}{the null hypothesis. Default vector is \eqn{(0,0)'} for a bivariate regression.}

\item{level}{the level of the test -i.e. \eqn{\alpha}. Default value is 0.05.}

\item{p}{the success probability of the binomial distribution for each trial. Default value is 0.5.}

\item{Dpndnt}{includes HAC-corrected signed statistic for serial (nonlinear) dependent data when set to TRUE.}

\item{B}{number of iterations for simulating the distribution of the test statistic under the null hypothesis.}

\item{bwidth}{the bandwidth for the Bartlett Kernell estimator.}
}
\description{
This function provides the nonparametric signed and HAC-corrected signed statstics proposed by \insertCite{coudin2009finite;textual}{Rdpack}.
The tests are distribution-free, exact and valid in the presence of multiple regressors, and errors exhibiting serial nonlinear dependence and heterogeneous volatility.
The distribution of the test statistic is simulated under the null hypothesis with sufficient number of replications and the critical values are derived respectively.
Unlike \insertCite{coudin2009finite;textual}{Rdpack}, this function is only limited to finite sample distributions. As such, the assumption of strict exogeneity must hold for the tests to be valid in small samples.
The R code for estimating the HAC covariance matrix has been retrieved from the paper by \insertCite{heberle2017fast;textual}{Rdpack}.
}
\examples{
SB_Dep(y,x,null=c(0,0),bwdith=2)
}
\references{
\insertAllCited{}
}
\keyword{HAC.}
\keyword{Nonparametric,}
\keyword{dependent,}
\keyword{exact,}
\keyword{inference,}
\keyword{signs,}
